Made changes to user interface
I took away the next button as the user was able to eventually select the correct answer so I made sure that once they answer its there final answer. I also added a green colour to the answer feedback for a correct answer and a red color for a wrong answer. I  felt that red was too strong a colour on the eyes when reading the introduction text, I wanted to see things from the users persepctive it wasnt good on the eyes so I opted for a vibrant orange which gives off a more energtic vibe rather then impending doom like the red color.